+++ src/main/java/com/revature/technology/repositories/ReimbStatusRepository.java
@@ -1,4 +1,4 @@
-package com.revature.technology.repos;
+package com.revature.technology.repositories;
 
 import com.revature.technology.models.ReimbursementStatus;
 import org.springframework.data.repository.CrudRepository;
