+++ src/main/java/com/revature/technology/repositories/ReimbTypeRepository.java
@@ -1,6 +1,5 @@
-package com.revature.technology.repos;
+package com.revature.technology.repositories;
 
-import com.revature.technology.models.ReimbursementStatus;
 import com.revature.technology.models.ReimbursementType;
 import org.springframework.data.repository.CrudRepository;
