+++ src/main/java/com/revature/technology/repositories/UserRepository.java
@@ -1,4 +1,4 @@
-package com.revature.technology.repos;
+package com.revature.technology.repositories;
 
 
 import com.revature.technology.models.User;
